bot/pyproject.toml
[tool.poetry]
name = "bot"
version = "1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.8"
"discord.py" = "^1.4.1"
python-dotenv = "^0.14.0"
colorama = "^0.4.3"
boto3 = "^1.15.1"
requests = "^2.24.0"
requests-oauthlib = "^1.3.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
